% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-df_sample_ids.R
\docType{data}
\name{df_sample_ids}
\alias{df_sample_ids}
\title{1% sample of Session ID values.}
\format{A \code{data.frame} object of one column and approximately 90,000 rows.}
\usage{
df_sample_ids
}
\description{
A \code{data.frame} object of Session IDs after sampling.
}
\details{
\describe{
    \item{Session_ID}{A sequence of integers, represented as a string, used to uniquely identify a session.}
}
}
\examples{
df_sample_ids
}

