% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sample_buys.R
\docType{data}
\name{sample_buys}
\alias{sample_buys}
\title{A sample of purchase data.}
\format{A \code{data.frame} object of five columns and approximately 12,000 rows.}
\usage{
sample_buys
}
\description{
A \code{data.frame} object of purchase data.
}
\details{
\describe{
    \item{Session_ID}{A sequence of integers, represented as a string, used to uniquely identify a session.}
    \item{Timestamp}{A timestamp of YYYY-MM-DD hh:mm:ss format indicating when a product was viewed.}
    \item{Item_ID}{A sequence of integers, represented as a string, used to uniquely identify a product.}
    \item{Price}{The price of the item.}
    \item{Quantity}{The quantity purchased.}
}
}
\examples{
sample_buys
}

