% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sample_clicks.R
\docType{data}
\name{sample_clicks}
\alias{sample_clicks}
\title{A sample of clickstream data.}
\format{A \code{data.frame} object of four columns and approximately 330,000 rows.}
\usage{
sample_clicks
}
\description{
A \code{data.frame} object of clickstream data, specifically product views.
}
\details{
\describe{
    \item{Session_ID}{A sequence of integers, represented as a string, used to uniquely identify a session.}
    \item{Timestamp}{A timestamp of YYYY-MM-DD hh:mm:ss format indicating when a product was viewed.}
    \item{Item_ID}{A sequence of integers, represented as a string, used to uniquely identify a product.}
    \item{Category}{A string used to identify the corresponding category of an item.}
}
}
\examples{
sample_clicks
}

